Fix sound issues with Safari on iOS14 when using M4A audio files
The newest Safari version doesn't return "maybe" for canPlayType('audio/mp4'), but instead just returns an empty string, indicating no support for this MIME type whatsoever. When we specify an exact codec (mp4a.40.2), canPlayType() reports the expected "probably".

I would consider this a bug in Safari, but don't have the patience to wait for a fix from Apple.